backend/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.18
django-cors-headers==4.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.3
typing-extensions==4.12.2
